prep cli and dependencies for implementation
