+++ src/cli.rs
@@ -1,6 +1,6 @@
-use crate::steganography::StegMethod;
 use anyhow::{bail, Result};
 use std::path::PathBuf;
+use std::str::FromStr;
 use structopt::StructOpt;
 
 #[derive(StructOpt)]
@@ -30,6 +30,10 @@ pub struct Opt {
     #[structopt(short, long, default_value = "lsb")]
     pub method: StegMethod,
 
+    /// Method for bit distribution (sequential, linear, random)
+    #[structopt(long, default_value = "sequential")]
+    pub distribution: BitDistribution,
+
     /// Seed for random significant bit encoding
     #[structopt(short, long, required_if("method", "rsb"))]
     pub seed: Option<String>,
@@ -37,6 +41,10 @@ pub struct Opt {
     /// Maximum bit to possible modify (1-4)
     #[structopt(short = "N", long, required_if("method", "rsb"))]
     pub max_bit: Option<u8>,
+
+    /// Seed for random bit distribution
+    #[structopt(long, required_if("distribution", "random"))]
+    pub distribution_seed: Option<String>,
 
     /// Output file, stdout if not present
     #[structopt(short, long, parse(from_os_str))]
@@ -59,5 +67,54 @@ impl Opt {
             }
         }
         Ok(())
+    }
+}
+
+/// Supported steganography encoding algorithms
+#[derive(StructOpt, Debug)]
+pub enum StegMethod {
+    /// Least significant bit encoding
+    ///
+    /// With a binary message, each bit of the message is encoded
+    /// into the least significant bit of each RGB byte of each pixel.
+    LeastSignificantBit,
+    /// Random significant bit encoding
+    ///
+    /// With a binary message, each bit of the message is encoded
+    /// randomly into one of the `n` least significant bits of each RGB byte of each pixel.
+    RandomSignificantBit,
+}
+
+impl FromStr for StegMethod {
+    type Err = String;
+    fn from_str(method: &str) -> Result<Self, Self::Err> {
+        match method {
+            "lsb" => Ok(Self::LeastSignificantBit),
+            "rsb" => Ok(Self::RandomSignificantBit),
+            other => Err(format!("unknown encoding method: {}", other)),
+        }
+    }
+}
+
+/// Supported bit encoding bit distribution methods
+#[derive(StructOpt, Debug)]
+pub enum BitDistribution {
+    /// Encode bits sequentially into the image starting from top-left
+    Sequential,
+    /// Evenly space out the bits in the image so not all packed into top-left
+    Linear,
+    /// Based on a random-seed, encode each bit into a random pixel and random colour channel
+    Random,
+}
+
+impl FromStr for BitDistribution {
+    type Err = String;
+    fn from_str(method: &str) -> Result<Self, Self::Err> {
+        match method {
+            "sequential" => Ok(Self::Sequential),
+            "linear" => Ok(Self::Linear),
+            "random" => Ok(Self::Random),
+            other => Err(format!("unknown bit distribution {}", other)),
+        }
     }
 }
